requirements.txt
opencv-python>=4.5
PyQt6>=6.4
PyQt6-Qt6>=6.4
PyQt6-QtMultimedia>=6.4
onvif-zeep==0.2.12
pymavlink>=2.4
numpy>=1.21
pyserial>=3.5
cryptography>=38.0
requests>=2.28